Change MCSample and SampleStack. Default copy construction of Walker is used to ensure memory consistency.
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5371 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/QMCWaveFunctions/EinsplineSetBuilderCommon.cpp
@@ -753,7 +753,7 @@ namespace qmcplusplus {
       // First make sure we have enough points
       if (superSets[si].size() != numTwistsNeeded) {
 	fprintf (stderr, "Super twist %d should own %d k-points, but owns %d.\n",
-		 si, numTwistsNeeded, superSets[si].size());
+		 si, numTwistsNeeded, static_cast<int>(superSets[si].size()));
 	abort();
       }
       // Now, make sure they are all distinct
